scaler: don't shift a clamped channel past the sign bit
clamp_8bit() returns uint8_t, which promotes to int before the shift,
so the alpha channel of the scalar vertical store overflows int for
any value above 127:

  scaler_int.c:225:32: runtime error: left shift of 138 by 24 places
  cannot be represented in type 'int'

Shift as uint32_t instead.  Every target we build for produces the
same store from either form, so this is UB on paper only - but it is
UB that fires on ordinary opaque pixels, and so buries the sanitizer
log of anything else in this file under a finding per output pixel.

Found by an ASan+UBSan sweep of the scaler; predates the current
work.  Output verified unchanged across an 11-case size matrix
(upscale, downscale, odd dimensions, unscaled).

+++ libretro-common/gfx/scaler/scaler_int.c
@@ -138,10 +138,10 @@ void scaler_argb8888_vert(const struct scaler_ctx *ctx, void *output_, int strid
          res_b           >>= (7 - 2 - 2);
 
          output[w]         =
-            (clamp_8bit(res_a) << 24) |
-            (clamp_8bit(res_r) << 16) |
-            (clamp_8bit(res_g) << 8)  |
-            (clamp_8bit(res_b) << 0);
+            ((uint32_t)clamp_8bit(res_a) << 24) |
+            ((uint32_t)clamp_8bit(res_r) << 16) |
+            ((uint32_t)clamp_8bit(res_g) << 8)  |
+            ((uint32_t)clamp_8bit(res_b) << 0);
 #endif
       }
    }
